move compress fns into streaming
