+++ src/deltas.ts
@@ -18,56 +18,6 @@ import {
   type StreamMessage,
 } from "./validators.js";
 
-/**
- * Compressing parts when streaming to save bandwidth in deltas.
- */
-
-export function compressUIMessageChunks(
-  parts: UIMessageChunk[],
-): UIMessageChunk[] {
-  const compressed: UIMessageChunk[] = [];
-  for (const part of parts) {
-    const last = compressed.at(-1);
-    if (part.type === "text-delta" || part.type === "reasoning-delta") {
-      if (last?.type === part.type && part.id === last.id) {
-        last.delta += part.delta;
-      } else {
-        compressed.push(part);
-      }
-    } else {
-      compressed.push(part);
-    }
-  }
-  return compressed;
-}
-
-export function compressTextStreamParts(
-  parts: TextStreamPart<ToolSet>[],
-): TextStreamPart<ToolSet>[] {
-  const compressed: TextStreamPart<ToolSet>[] = [];
-  for (const part of parts) {
-    const last = compressed.at(-1);
-    if (part.type === "text-delta" || part.type === "reasoning-delta") {
-      if (last?.type === part.type && part.id === last.id) {
-        last.text += part.text;
-      } else {
-        compressed.push(part);
-      }
-    } else {
-      if (part.type === "file") {
-        compressed.push({
-          type: "file",
-          file: {
-            ...part.file,
-            uint8Array: undefined as unknown as Uint8Array,
-          },
-        });
-      }
-      compressed.push(part);
-    }
-  }
-  return compressed;
-}
 
 export function blankUIMessage<METADATA = unknown>(
   streamMessage: StreamMessage & { metadata?: METADATA },
